+++ examples/complete/outputs.tf
@@ -12,4 +12,3 @@ output "test_log_analytics_workspace_resource_id" {
   description = "value of the resource ID for the Log Analytics Workspace."
   value       = module.management.log_analytics_workspace_resource_id
 }
-